+++ product/src/main/java/com/study/product/controller/ProductController.java
@@ -1,8 +1,11 @@
 package com.study.product.controller;
 
 import com.study.product.controller.docs.ProductControllerDocs;
+import com.study.product.model.request.DiscountSaleProductRequestDTO;
+import com.study.product.model.request.ProductOrderRequestDTO;
 import com.study.product.model.request.ProductRequestDTO;
 import com.study.product.model.request.SearchConditionDTO;
+import com.study.product.model.response.ProductOrderResponseDTO;
 import com.study.product.model.response.ProductSearchResultDTO;
 import com.study.product.model.response.Response;
 import com.study.product.service.ProductService;
@@ -11,6 +14,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/product")
@@ -18,12 +23,19 @@ public class ProductController implements ProductControllerDocs {
 
 	private final ProductService productService;
 
-	//todo: 권한 추가하기
 	@PostMapping
 	public Response<Long> createProduct(@Valid @RequestBody ProductRequestDTO request) {
 
 		return Response.<Long>builder()
 				.data(productService.createProduct(request))
+				.build();
+	}
+
+	@PostMapping("/set-up")
+	public Response<Void> setUpDiscountSaleProduct(@Valid @RequestBody DiscountSaleProductRequestDTO request) {
+
+		productService.setUpDiscountSaleProduct(request);
+		return Response.<Void>builder()
 				.build();
 	}
 
@@ -35,6 +47,12 @@ public class ProductController implements ProductControllerDocs {
 				.build();
 	}
 
+	@PatchMapping("/order")
+	public List<ProductOrderResponseDTO> modifyProductStock(@Valid @RequestBody List<ProductOrderRequestDTO> request) {
+
+		return productService.modifyProductStock(request);
+	}
+
 	@GetMapping
 	public Response<ProductSearchResultDTO> getCurrentSaleProductList(Pageable pageable,
 	                                                                  SearchConditionDTO searchCondition) {
@@ -42,5 +60,11 @@ public class ProductController implements ProductControllerDocs {
 		return Response.<ProductSearchResultDTO>builder()
 				.data(productService.getCurrentSaleProductList(pageable, searchCondition))
 				.build();
+	}
+
+	@GetMapping("/order")
+	public List<ProductOrderResponseDTO> getProductOrderList(@Valid @RequestParam List<Long> productIdList) {
+
+		return productService.getProductOrderList(productIdList);
 	}
 }
