+++ product/src/main/java/com/study/product/controller/docs/ProductControllerDocs.java
@@ -1,7 +1,9 @@
 package com.study.product.controller.docs;
 
+import com.study.product.model.request.ProductOrderRequestDTO;
 import com.study.product.model.request.ProductRequestDTO;
 import com.study.product.model.request.SearchConditionDTO;
+import com.study.product.model.response.ProductOrderResponseDTO;
 import com.study.product.model.response.ProductSearchResultDTO;
 import com.study.product.model.response.Response;
 import io.swagger.v3.oas.annotations.Operation;
@@ -13,6 +15,8 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @Tag(name = "Product", description = "제품 등록,수정 및 조회 등의 사용자 API")
 public interface ProductControllerDocs {
@@ -41,5 +45,21 @@ public interface ProductControllerDocs {
 			Pageable pageable,
 			SearchConditionDTO searchCondition
 	);
+
+	@Operation(summary = "제품 주문", description = "제품을 주문하는 API 입니다.")
+	@ApiResponses(value = {
+			@ApiResponse(responseCode = "200", description = "주문 성공", content = @Content(schema = @Schema(implementation = Response.class))),
+			@ApiResponse(responseCode = "400", description = "재고가 부족합니다.", content = @Content(schema = @Schema(implementation = Response.class)))
+	})
+	@PatchMapping("/product/order")
+	List<ProductOrderResponseDTO> modifyProductStock(@Valid @RequestBody List<ProductOrderRequestDTO> request);
+
+	@Operation(summary = "제품 주문 전단계", description = "제품 주문 전단계에서 주문 가능 여부를 확인하고, 재고와 가격을 확인하는 API 입니다.")
+	@ApiResponses(value = {
+			@ApiResponse(responseCode = "200", description = "주문 가능 여부 확인 성공", content = @Content(schema = @Schema(implementation = Response.class))),
+			@ApiResponse(responseCode = "400", description = "주문 불가능한 제품이 포함되어 있습니다.", content = @Content(schema = @Schema(implementation = Response.class)))
+	})
+	@GetMapping("/product/order")
+	List<ProductOrderResponseDTO> getProductOrderList(@Valid @RequestParam List<Long> productIdList);
 
 }
